pyproject.toml
[project]
name = "marketnews"
version = "0.1.0"
description = "Daily monitor of technology market news and sentiment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "mysql-connector-python>=9.1.0",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
]
